requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==18.2.0
backcall==0.1.0
coverage==4.5.2
decorator==4.3.2
Django==2.1.7
djangorestframework==3.9.1
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
more-itertools==5.0.0
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.1
prompt-toolkit==2.0.8
ptyprocess==0.6.0
py==1.7.0
Pygments==2.3.1
pytest==4.2.0
pytest-cov==2.6.1
pytest-django==3.4.7
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
yapf==0.26.0
